move package + docs
