+++ plugins/source/aws/resources/services/s3control/access_points_mock_test.go
@@ -1,4 +1,4 @@
-package s3
+package s3control
 
 import (
 	"testing"
